Tweaks and changes to x32/extension
