--- types/X32.d.ts
@@ -1,7 +1,0 @@
-export namespace X32 {
-  interface Config {
-    enable: boolean;
-    ip: string;
-    localPort: number;
-  }
-}
