+++ types/index.d.ts
@@ -4,4 +4,3 @@ export * from './OBS';
 export * from './RabbitMQ';
 export * from './Restream';
 export * from './Tracker';
-export * from './X32';
