+++ x32/extension/.eslintignore
@@ -1,1 +1,3 @@
 .eslintrc.js
+dist/**/*
+node_modules/**/*
