+++ x32/extension/.eslintrc.js
@@ -1,6 +1,36 @@
 module.exports = {
-  extends: ['../../.eslintrc.extension.js'],
+  root: true,
+  env: {
+    node: true,
+  },
+  parser: '@typescript-eslint/parser',
   parserOptions: {
     project: 'tsconfig.json',
+  },
+  plugins: [
+    '@typescript-eslint',
+  ],
+  extends: [
+    'airbnb-typescript/base',
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:import/typescript',
+  ],
+  settings: {
+    'import/resolver': {
+      typescript: {
+        // intentionally left blank
+      },
+    },
+  },
+  rules: {
+    'import/extensions': ['error', 'ignorePackages', {
+      js: 'never',
+      jsx: 'never',
+      ts: 'never',
+      tsx: 'never',
+    }],
+    'max-len': ['error', { code: 100, ignorePattern: '^import\\s.+\\sfrom\\s.+;$' }],
+    '@typescript-eslint/lines-between-class-members': 'off',
   },
 };
