--- x32/extension/.gitignore
@@ -1,2 +1,0 @@
-# Built files
-#/dist
