+++ x32/extension/package.json
@@ -6,5 +6,8 @@
     "autofix": "node ../../node_modules/eslint/bin/eslint.js --fix --ext .ts src",
     "build": "node ../../node_modules/typescript/bin/tsc -b tsconfig.json",
     "watch": "node ../../node_modules/typescript/bin/tsc -b tsconfig.json -w"
+  },
+  "dependencies": {
+    "osc": "^2.4.1"
   }
 }
